+++ apps/api/src/user/dto/onboarding-answers/onboarding-answers.ts
@@ -1,0 +1,38 @@
+import { IsOptional, IsString } from 'class-validator'
+
+export class OnboardingAnswersDto {
+  @IsString()
+  name: string
+
+  @IsString()
+  @IsOptional()
+  profilePictureUrl?: string
+
+  @IsString()
+  @IsOptional()
+  role?: string
+
+  @IsString()
+  @IsOptional()
+  industry?: string
+
+  @IsString()
+  @IsOptional()
+  teamSize?: string
+
+  @IsString()
+  @IsOptional()
+  productStage?: string
+
+  @IsString()
+  @IsOptional()
+  useCase?: string
+
+  @IsString()
+  @IsOptional()
+  heardFrom?: string
+
+  @IsString()
+  @IsOptional()
+  referralCode?: string
+}