+++ apps/platform/src/components/shared/navbar/profile-menu.tsx
@@ -67,7 +67,7 @@ export default function ProfileMenu() {
             )}
             <DropdownSVG />
           </DropdownMenuTrigger>
-          <DropdownMenuContent>
+          <DropdownMenuContent align="end">
             <DropdownMenuLabel>My Account</DropdownMenuLabel>
             <DropdownMenuSeparator />
             <Link href="/settings?tab=profile">
@@ -81,9 +81,6 @@ export default function ProfileMenu() {
             </Link>
             <Link href="/teams">
               <DropdownMenuItem>Team</DropdownMenuItem>
-            </Link>
-            <Link href="/settings?tab=profile">
-              <DropdownMenuItem>User Settings</DropdownMenuItem>
             </Link>
 
             <DropdownMenuSeparator />